boostcoach/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
prisma
psycopg2-binary